Rename original JFLAP translator file to be clearer
