+++ packages/jflap-translator/src/convertJFLAPToAutomatarium.ts
