+++ packages/jflap-translator/src/index.ts
@@ -1,2 +1,2 @@
-export { convertJFLAPXML, convertJFLAPProject } from './convertJFLAP'
+export { convertJFLAPXML, convertJFLAPProject } from './convertJFLAPToAutomatarium'
 export { convertAutomatariumToJFLAP } from './convertAutomatariumToJFLAP'
